jira-ingester/requirements.txt
functions-framework==3.8.1
google-cloud-pubsub==2.21.1
Flask==3.0.3
